requirements.txt
pygame>=2.5
pyyaml>=6.0
numpy>=1.24
torch>=2.0
pillow>=9.0
